Add Dockerfile
What?
====

Here we add a file that is used by [Docker][1] to build and run a
containerised version of our app. At the moment this is a fairly
basic setup to get our up and running.

Why?
===

We will be using Heroku to deploy our app to, which we will trigger
using GitHub Actions.

Heroku requires either a Procfile or Dockerfile to be present.
Deploying an app onto containerised infrastructure is quite common,
and Docker is the most widely used tool for this job. As such if
we wanted to switch from Heroku later, it may be easier to find a
new platform that also supports dockerised deploys than finding
options that read from Procfiles in a way similar to Heroku.

Docker may also offer us some development benefits. We can run
the application within docker whilst we make changes, reducing
points of difference between our development Env and production
Env.

As we expect development to take place on different machines
with different operating systems, a containerised environment may
also help prevent development anomolies occuring by providing some
consistency to how dependencies are obtained and run between
machines.

How?
===

We start by defining a base image on line 1. This is an image
someone else has written, often with a lot of other setup that gets
us to a known point. In this case we are using the ruby:3.0.0 image
which is both widely used and well maintained. With this line alone
we should end up with a working linux system with all the
dependencies required to run a ruby program.

We then use some linux package manager commands to bring in a few
system level rails dependencies.

Then make a folder in our containerised world called /app.
Copy across the Gemfile and Gemfile.lock which define our rails
dependencies and use bundle to install them.

Finally we copy across everything else into /app, expose port 3001
and run the rails server (rails s) command to start the server.

[1]: https://www.docker.com/